+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 ### Minor
 
+- Text [Breaking]: Removes deprecated size=xl (#729)
+
+Run codemod:
+
+`cd gestalt; yarn run codemod --parser=flow -t=packages/gestalt-codemods/1.14.0-1.15.0/remove-text-size-xl.js ~/code/repo`
+
 ### Patch
 
 </details>